feat: add an authetication option callback to the metrics route
this will be a new option to the middleware config. By default, no
authentication is required.

If a function is provided in the options, the function will be called with the
`req` object, and should return thuthy for sucessfull authentications, or falsy
otherwise. In case authentication callback returns falsy or throws, the `next`
callback is called without errors, and will likely return a `404` status code
for the requester, to avoid hinting at the existance of the metrics route