feat(option):  resetOnScrape boolean option
body:
if set to true resetOnScrape will call resetMetrics on the default metric registry. This has the advantage that it will stop memo
ry piling up in the metrics themselves. promethues scrape performance degrades if the metrics return a lot of data.

Prometheus keeps deta that it scrapes in a time series database, so we don't need it hanging about, unless you are using long liv
ed counters.

resetOnScrape defaults to false